requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
einops==0.8.2
fastapi==0.134.0
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.12.0
gradio==6.8.0
gradio_client==2.2.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
semantic-version==2.10.0
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
spandrel==0.4.2
starlette==0.52.1
sympy==1.14.0
tkinterdnd2-universal==1.7.3
tomlkit==0.13.3
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
